Update dutch translation Iota wallet.
Hello Dominique, I saw a misspelled dutch word in the wallet, so I mentioned on the dev's slack channel. And someone said I could check it myself and gave me the file. So I changed the words and found a few other misspelled words and some capitals in the wrong place. Of course I left the English words which are so common these days in our language untouched.
If you want can help with translating or other stuff I can do to help the Iota project. (Edwin, living near Amsterdam Netherlands)